add optimize methods

+++ src/BoxInterface.java
@@ -2,7 +2,7 @@ public interface BoxInterface {
     void addSweetness(Sweetness sweetness);
     Sweetness remove(int index);
     Sweetness removeLast();
-    double getWeightBox();
+    int getWeightBox();
     int getPriceBox();
     void showAllInformation();
 }
